ATACTracks objects

ATACTracks is a shallow object holding ATAC data over cells/metacells. Minimally it should include a vector of track names which hold the names of the tracks with the ATAC data. McTracks extend the ATACTracks object by adding metadata and additional slots.

Slots

tracks

vector of track names with the ATAC data.

total_cov

vector with total coverage of the tracks.

marginal_track

name of a track with marginal coverage.

id

an identifier for the object, e.g. "pbmc".

description

description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"

genome

genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"

metadata

data frame with a column called 'metacell' and additional metacell annotations. The constructor can also include or the name of a delimited file which contains such annotations.

resolution

the resolution of the tracks.

window_size

The size of the window used to smooth the counts in each track. If no smoothing was done - this should be equal to resolution/2