Calculate gene markers of metacells gene expression

The function first removes genes without sufficient expression in any metacell minimal_max_log_fraction, or without sufficient fold change over the median (minimal_relative_expression), and then ranks the fold change of genes within each metacell. The markers are then the genes with the highest rank, up to n_genes genes.

Usage

get_rna_markers(
  atac_mc,
  n_genes = 100,
  minimal_max_log_fraction = -13,
  minimal_relative_log_fraction = 2,
  fold_change_reg = 0.1,
  genes = NULL,
  rm_zeros = TRUE,
  epsilon = 0.00001
)

Arguments

atac_mc

a McPeaks/McTracks object with RNA expression (using add_mc_rna)

n_genes

maximal number of genes to return. Default: 100

minimal_max_log_fraction

take only genes with at least one value (in log fraction units - normalized egc) above this threshold

minimal_relative_log_fraction

take only genes with at least one value with relative log fraction (mc_fp) above this this value

fold_change_reg

regularization factor for the fold change calculation (fold_change would be changed to fold_change = fold_change + fold_change_reg)

genes

list of genes to match. Default (NULL): all genes

rm_zeros

remove genes with no RNA expression in any metacell. Default: TRUE

epsilon

regularization factor added to the log normalized expression

Examples

if (FALSE) {
get_rna_markers(atac_mc)
}