Create an McTracks object

Create an McTracks object either from a list of tracks (tracks parameter) or from a track prefix (track_prefix parameter). The tracks are assumed to be in the form of ".+_mcmetacell_number", if this is not the case, the metacells parameter can be used to specify the metacell number for each track.

Usage

mct_create(
  genome,
  tracks = NULL,
  track_prefix = NULL,
  metacells = NULL,
  marginal_track = NULL,
  id = NULL,
  description = NULL,
  path = NULL,
  metadata = NULL,
  resolution = NULL,
  window_size = resolution/2,
  order = NULL
)

Arguments

genome

genome assembly of the tracks. e.g. "hg38", "hg19", "mm9", "mm10"

tracks

a vector of tracks to include in the object.

track_prefix

prefix for tracks (instead of explicit track names at the tracks parameter). Ignored if tracks is not NULL. If a track named "track_prefix.marginal" exists, it will be used as the marginal track.

metacells

a vector of metacells to include in the object (optional). If not provided, the metacells are inferred from the track names using the pattern "track_prefix.mc\d*$". In such a case, the tracks are ordered according to the metacell number.

marginal_track

a track to use as the marginal track (optional). If not provided, and track_prefix was given, the track named "track_prefix.marginal" will be used if it exists.

id

an identifier for the object, e.g. "pbmc".

description

description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"

path

path to a directory containing the raw data. Usually inherited from the counts object.

metadata

data frame with a column called 'metacell' and additional metacell annotations, or the name of a delimited file which contains such annotations.

resolution

the resolution of the tracks.

window_size

The size of the window used to smooth the counts in each track. If no smoothing was done - this should be equal to resolution / 2.

order

the order of the tracks in the object. If not provided, the tracks are ordered according to the metacell number.

Value

a McTracks object.

Examples

if (FALSE) {
mct_create(track_prefix = "pbmc_mc", genome = "hg38", id = "pbmc", description = "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)")
}