Write single cell counts from a 10x fragments file

This function writes an ScCounts object from a "fragments.tsv.gz" which is an output of the 10x pipeline ("Barcoded and aligned fragment file").

Usage

scc_from_fragments(
  fragments_file,
  out_dir,
  cell_names,
  cell_suffix = NULL,
  genome = NULL,
  bin_size = 50000000,
  overwrite = FALSE,
  id = "",
  description = "",
  verbose = FALSE,
  tabix_bin = "tabix",
  chromosomes = gintervals.all()$chrom
)

write_sc_counts_from_fragments(
  fragments_file,
  out_dir,
  cell_names,
  cell_suffix = NULL,
  genome = NULL,
  bin_size = 50000000,
  overwrite = FALSE,
  id = "",
  description = "",
  verbose = FALSE,
  tabix_bin = "tabix",
  chromosomes = gintervals.all()$chrom
)

Arguments

fragments_file

path to fragments file. Note that in order to use tabix, the file must be compressed with gzip and have a ".gz" extension.

out_dir

output directory.

cell_names

a vector with the cell names or an ScPeaks object

cell_suffix

a suffix to add to the cell names (optional). This is useful when you want to merge multiple batches of cells. The suffix would be added in the following way: paste0(cell_name, "-", cell_suffix).

bin_size

Size of the genomic bins to use (in bp). Each chromsome will be chunked into bins with size which is smaller than this value. Default is 50Mb.

overwrite

overwrite existing directory (optional)

id

an identifier for the object, e.g. "pbmc".

description

description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"

verbose

verbose output (optional)

tabix_bin

path to the tabix binary (optional)

chromosomes

process only specific chromosomes (optional).

Value

None

Examples

if (FALSE) {
scc_from_fragments("pbmc_data/fragments.tsv.gz", "pbmc_reads", cell_names = atac_sc)
scc_from_fragments("pbmc_data/fragments.tsv.gz", "pbmc_reads", cell_names = atac_sc, chromosomes = paste0("chr", c(1:22, "X", "Y")))
}